refactor(qbytes): specialize activation and weights
This is a huge refactoring but actually quite simple in design: instead
of having a single QBytesTensor class for activations and weights, we
now have two classes: WeightQBytesTensor and ActivationQBytesTensor.
Each class has its own dispatch handler for operations, but only the
WeightQBytesTensor class implements a dispatch at the function level.
This removes a lot of ambiguity when dispatching for functions that can
accept both quantized weights and activations at the same time: if no
suitable dispatch is found for the weights, then both tensors are simply
dequantized.
For operations, there is an uncertainty about which dispatch handler will
be called first, but since both dispatch handlers only implement
operations that accept a single type of QTensor, it will led again to a
dequantization.
