+++ optimum/quanto/tensor/activations/quantization.py
@@ -14,8 +14,8 @@
 
 import torch
 
-from .qbytes import QBytesTensor
-from .qtype import qtype
+from ..qtype import qtype
+from .qbytes import ActivationQBytesTensor
 
 
 __all__ = ["quantize_activation"]
@@ -36,4 +36,4 @@ def quantize_activation(t: torch.Tensor, qtype: qtype, scale: torch.Tensor):
     """
     if scale.numel() != 1:
         raise ValueError("Parameter scale must be a scalar because activations can only be quantized per-tensor")
-    return QBytesTensor.quantize(t, qtype, None, scale)
+    return ActivationQBytesTensor.quantize(t, qtype, scale)
