+++ test/nn/test_qconv2d.py
@@ -16,7 +16,7 @@ import pytest
 import torch
 from helpers import assert_similar, random_qactivation, random_tensor
 
-from optimum.quanto import Calibration, QBytesTensor, qfloat8_e4m3fn, qfloat8_e5m2, qint4, qint8
+from optimum.quanto import ActivationQBytesTensor, Calibration, qfloat8_e4m3fn, qfloat8_e5m2, qint4, qint8
 from optimum.quanto.nn import QConv2d
 
 
@@ -29,7 +29,7 @@ def _test_quantize_conv2d(batch_size, img_shape, out_channels, use_bias, weights
     with torch.no_grad(), Calibration():
         qout = qconv2d(inputs)
     if activations is not None:
-        assert isinstance(qout, QBytesTensor)
+        assert isinstance(qout, ActivationQBytesTensor)
         assert qout.qtype == activations
     # Align weights with quantized linear weights for comparison
     conv2d.weight = torch.nn.Parameter(qconv2d.qweight.dequantize())
