+++ test/nn/test_qlinear.py
@@ -20,9 +20,9 @@ import torch
 from helpers import assert_similar, random_qactivation, random_tensor
 
 from optimum.quanto import (
+    ActivationQBytesTensor,
     Calibration,
     QBitsTensor,
-    QBytesTensor,
     absmax_scale,
     qfloat8,
     qfloat8_e4m3fn,
@@ -49,7 +49,7 @@ def _test_quantize_linear(batch_size, tokens, embeddings, use_bias, weights, act
     with torch.no_grad(), context():
         qout = qlinear(inputs if activations is None else qinputs)
     if activations is not None:
-        assert isinstance(qout, QBytesTensor)
+        assert isinstance(qout, ActivationQBytesTensor)
         assert qout.qtype == activations
     # Align linear weights with quantized linear weights for comparison
     linear.weight = torch.nn.Parameter(qlinear.qweight.dequantize())
