+++ test/tensor/activations/test_activations_dispatch.py
@@ -16,7 +16,7 @@ import pytest
 import torch
 from helpers import assert_similar, random_qactivation, random_tensor
 
-from optimum.quanto import QBytesTensor, quantize_activation
+from optimum.quanto import ActivationQBytesTensor, quantize_activation
 
 
 @pytest.mark.parametrize("input_shape", [(10,), (1, 10), (10, 32, 32)])
@@ -26,11 +26,11 @@ def test_qactivation_mul_scalar(input_shape, scalar, device):
     if isinstance(scalar, torch.Tensor):
         scalar = scalar.to(device)
     qprod = qa * scalar
-    assert isinstance(qprod, QBytesTensor)
+    assert isinstance(qprod, ActivationQBytesTensor)
     prod = qa.dequantize() * scalar
     assert_similar(prod, qprod)
     qprod = scalar * qa
-    assert isinstance(qprod, QBytesTensor)
+    assert isinstance(qprod, ActivationQBytesTensor)
     prod = scalar * qa.dequantize()
     assert_similar(prod, qprod)
 
@@ -40,7 +40,7 @@ def test_qactivation_mul_scalar(input_shape, scalar, device):
 def test_qactivation_relu(batch_size, tokens, embeddings, device):
     qinputs = random_qactivation((batch_size,) + (tokens, embeddings), dtype=torch.float32).to(device)
     qout = torch.nn.functional.relu(qinputs)
-    assert isinstance(qout, QBytesTensor)
+    assert isinstance(qout, ActivationQBytesTensor)
     assert torch.equal(qout._data, torch.maximum(qinputs._data, torch.zeros((1,)).to(device)))
 
 
@@ -49,7 +49,7 @@ def test_qactivation_relu(batch_size, tokens, embeddings, device):
 def test_qactivation_softmax(batch_size, tokens, embeddings, device):
     qinputs = random_qactivation((batch_size,) + (tokens, embeddings), dtype=torch.float32).to(device)
     qout = torch.nn.functional.softmax(qinputs, dim=-1)
-    assert isinstance(qout, QBytesTensor)
+    assert isinstance(qout, ActivationQBytesTensor)
     assert torch.min(qout.dequantize()) >= 0
     assert torch.max(qout.dequantize()) <= 1
 
@@ -58,7 +58,7 @@ def test_qactivation_softmax(batch_size, tokens, embeddings, device):
 def test_qactivation_view(input_shape, device):
     qinputs = random_qactivation(input_shape, dtype=torch.float32).to(device)
     qview = qinputs.view((1,) + input_shape)
-    assert isinstance(qview, QBytesTensor)
+    assert isinstance(qview, ActivationQBytesTensor)
 
 
 @pytest.mark.parametrize("input_shape", [(10,), (10, 32)])
@@ -68,7 +68,7 @@ def test_qactivation_cat(input_shape, device):
     # First, quantize other with the same scale
     qother = quantize_activation(other, qtype=qinputs.qtype, scale=qinputs._scale)
     qcat = torch.cat([qinputs, qother])
-    assert isinstance(qcat, QBytesTensor)
+    assert isinstance(qcat, ActivationQBytesTensor)
     assert_similar(torch.cat([qinputs.dequantize(), qother.dequantize()]), qcat)
